checkstyle fix attempt
