--- src/main/java/seedu/duke/Shelving/Shelves/RomanceShelves.java
@@ -1,10 +1,0 @@
-package seedu.duke.Shelving.Shelves;
-
-import seedu.duke.book.Book;
-import seedu.duke.exception.SectionFullException;
-
-public class RomanceShelves extends Shelves {
-    public RomanceShelves() {
-        super("R");
-    }
-}
