--- src/main/java/seedu/duke/Shelving/Shelves/SciFiShelves.java
@@ -1,11 +1,0 @@
-package seedu.duke.Shelving.Shelves;
-
-import seedu.duke.book.Book;
-import seedu.duke.exception.SectionFullException;
-
-public class SciFiShelves extends Shelves {
-
-    public SciFiShelves() {
-        super("SCIF");
-    }
-}
