+++ src/main/java/seedu/duke/shelving/shelves/ActionShelves.java
@@ -1,0 +1,7 @@
+package seedu.duke.shelving.shelves;
+
+public class ActionShelves extends Shelves {
+    public ActionShelves() {
+        super("AC");
+    }
+}
