+++ src/main/java/seedu/duke/shelving/shelves/AdventureShelves.java
@@ -1,0 +1,7 @@
+package seedu.duke.shelving.shelves;
+
+public class AdventureShelves extends Shelves {
+    public AdventureShelves() {
+        super("AD");
+    }
+}
