+++ src/main/java/seedu/duke/shelving/shelves/HorrorShelves.java
@@ -1,0 +1,7 @@
+package seedu.duke.shelving.shelves;
+
+public class HorrorShelves extends Shelves {
+    public HorrorShelves() {
+        super("H");
+    }
+}
