+++ src/main/java/seedu/duke/shelving/shelves/MysteryShelves.java
@@ -1,0 +1,7 @@
+package seedu.duke.shelving.shelves;
+
+public class MysteryShelves extends Shelves {
+    public MysteryShelves() {
+        super("MY");
+    }
+}
