+++ src/main/java/seedu/duke/shelving/shelves/NonFictionShelves.java
@@ -1,0 +1,7 @@
+package seedu.duke.shelving.shelves;
+
+public class NonFictionShelves extends Shelves{
+    public NonFictionShelves() {
+        super("NF");
+    }
+}
