+++ src/main/java/seedu/duke/shelving/shelves/RomanceShelves.java
@@ -1,0 +1,7 @@
+package seedu.duke.shelving.shelves;
+
+public class RomanceShelves extends Shelves {
+    public RomanceShelves() {
+        super("R");
+    }
+}
