+++ src/main/java/seedu/duke/shelving/shelves/SciFiShelves.java
@@ -1,0 +1,8 @@
+package seedu.duke.shelving.shelves;
+
+public class SciFiShelves extends Shelves {
+
+    public SciFiShelves() {
+        super("SCIF");
+    }
+}
